+++ R/globals.R
@@ -1,11 +1,7 @@
 # Declare global variables used in non-standard evaluation to satisfy R CMD check.
 utils::globalVariables(c(
-  "D",
-  "D_positive",
   "density",
   "grid",
-  "observation",
-  "score",
   "x",
   "y",
   "z"